Update patch set 9
Patch Set 9:

If the start() and stop() names are not mandatory, as they are called because of the annotations: PostConstruct, PreConstruct, you can change their names to whatever: initLifecycle/startLifecycle/infrastructureStart, and then change the current startWithException() to start().

On a different subject:
My approach (And I know others would disagree) would be to make the start and stop abstract and force the extendor to implement them. I think in the long run that this approach is easier for the user, because he can't even compile without "Add unimplemented methods", and then his "job" is more clear.

Patch-set: 9
Label: Verified=0