Update patch set 2
Patch Set 2:

Patch Set 2 has fixed the real problem of this change (the versions were not correct; back-patching NEW wor kis NOT as easy as a simple cherry-pick!).  

The current build failure () now appears to be unrelated to this change itself?  To me it looks like stable/boron is currently broken somewhere. I'll not investigate this problem further at this time, and may simply retry to build this in a day or two or so:

Execution default-bundle of goal org.apache.felix:maven-bundle-plugin:3.0.1:bundle failed. NullPointerException

11:54:18 installFeature(org.opendaylight.odlparent.featuretest.SingleFeatureTest)[repoUrl: file:/w/workspace/infrautils-distribution-check-boron/distribution/features-test/target/classes/features.xml, Feature: odl-integration-compatible-with-all 0.5.1-SNAPSHOT]  Time elapsed: 20.611 sec  <<< ERROR!
11:54:18 java.lang.IllegalStateException: Can't install feature odl-integration-compatible-with-all/0.5.1-SNAPSHOT: 	
11:54:18 Jar is not a bundle, no Bundle-SymbolicName mvn:org.opendaylight.openflowplugin.model/model-flow-base/0.3.1-SNAPSHOT

Patch-set: 2