Update patch set 1
Patch Set 1: Code-Review-1

Tx! Oh it's that easy.. nice.  This is an example to illustrate, not something to actually merge (yet), agreed?  Lemme -1 it just to avoid other infrautils commiters merging it by mistake.  If folks to want to build karaf4 dists, perhaps we could have parallel karaf4/ folders next to where we have karaf/, like we do for features?

Patch-set: 1
Label: Code-Review=-1