Update patch set 3
Patch Set 3:

Guy Sela has pointed out a potential problem with the current design, in some places that use the DSJC (such as netvirt), we sometimes get into collisions at mdsal level due to using job keys from different "levels". For example, in netvirt we process elan-interface objects and their static-mac-entries. In some places we use the elan interface name as key (thus we're locking on that key) and in other places we create jobs for each static-mac-entry (thus locking on them). The problem is that the job which is created for the elan-interface also processes static-mac-entries and we get collisions (the ones we tried to avoid by using the coordinator). A possible solution for this is to use a "hierarchical" way of representing jobs with a matching locking mechanism. Considering my last example, instead of locking on the static-mac-entry, we could lock on something like "elaninterfacename.staticmacentry" (. is the special delimiter here), and then the DSJC will first try to lock on the first part of the key (the elaninterfacename). What do you think?

Patch-set: 3
Label: Code-Review=0