Update patch set 3
Patch Set 3:

Faseela,

My general understanding from this spec is that we want to delay opening of various sockets such as OF/OVSDB/OpenStack until system is considered up and running. 
What I am missing most after reading it is who will these lower-layer services like OFP that have no notion of netvirt/genius etc. query the monitoring-service? Will they get the status of all registered services and decide to open the socket only when all are considered OPERATIONAL? What about services that are subject to configuration like BgpVpnManager?

It would be very helpful to add some concrete usage examples and also to explain if some consumers of the monitoring service will be implemented as part of this feature or this is just a general infrastructure and the responsibility to get the state and react to state changes has to be implemented per-project.

Patch-set: 3
Reviewer: Gerrit User 5802 <5802@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0