Update patch set 3
Patch Set 3:

This may not be a problem with djc and it is just about enhancing djc to support hierarchical jobs using its key. 
It is achieved in applications with enqueuing a job inside another job (example: InterfaceAddWorkerOnElan worker is on elanInstanceName key which again enqueues InterfaceAddWorkerOnElanInterface job using interfaceName at end of its run). 
If DJC implements this, we need to decide on a specific scenario. For example, while executing jobs at 4th hierarchical level, there should be jobs enqueued for levels < 4, how do we handle it ? Do we halt the jobs that are >=4 and execute the enqueued jobs ?

Patch-set: 3
Label: Code-Review=0