Update patch set 17
Patch Set 17:

FTR: Yakir tx for having taken my earlier review feedback into account.. from a pure code of view this is OK as-is with me now; no more objection from me to merge it..  I'm currently just hesitant to +1/+2 it because of the ongoing discussion in https://git.opendaylight.org/gerrit/#/c/51418/ ... so what we have here is a cleaned up version of what we have in genius today.  Whether this utility makes sense at all is what the discussion in c/51418 is all about now.

Patch-set: 17
Label: Verified=0