Update patch set 17
Patch Set 17:

I am fine with this being submitted, as long as we are going to to revisit this in the next release.

Notably the data races should not be common and point towards unsynchronized access to same data trees, which runs contrary to data-driven design.

Patch-set: 17
Reviewer: Gerrit User 649 <649@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0