Update patch set 7
Patch Set 7: Code-Review-1

You wouldn't want to cover this change with a unit test? With all the escaping and change from one util to another, I wouldn't be able to guarantee that this still does exactly the same thing (it probably does, as you know what you're doing, but if you have time for a test, it would make me give a +2 with more confidence on this one).

Patch-set: 7
Label: Code-Review=-1