Minor: remove MANIFEST.MF from .gitignore (and add .vscode/)
There used to be a build infrastructure problem with made the
maven-bundle-plugin generate the MANIFEST.MF into src/main/resources,
and this entry in .gitignore is probably from back then.  This was
wrong, and has been fixed since, and now MANIFEST.MF is generated into
target/ as it should - and so this can be removed now.

Change-Id: Ic9c8e62f19819f8639e6cd076a46bd7fc3acee59
Signed-off-by: Michael Vorburger <vorburger@redhat.com>