Update patch set 15
Patch Set 15:

Thanks for reviewing Robert!
The primary reason for not having any yang model is that, we don't intend to store these status in datastore. Also, northbound applications won't use any RPCs to query the status from status-and-diag. The status is exposed as MBean, which can be directly read by applications, and this avoids having tight dependancies of other projects on status-and-diag. Currently these API classes are only used by the status-and-diag shell bundle. Let me know your thoughts

Patch-set: 15
Label: Verified=0