Update patch set 3
Patch Set 3: Code-Review-1

I'm against merging this, which eliminates the Karaf 3 distribution that is useful for local testing of infrautils developments (I've used it last week to work on the Caches), until ODL overall has completely switched over to be Karaf 4 only (i.e. just before the Karaf 3 related artifacts in odlparent will have been removed). For the moment I get the impression that we live with both, and while I'm hoping to see only Karaf 4 everywhere soon, we're not there yet.  When we're ready to do that, this patch could go further and even completely remove the Karaf 3 features, not just the distribution.

Patch-set: 3
Label: Code-Review=-1