Update patch set 3
Patch Set 3:

(2 comments)

For me Executors is just fine. We can also narrow it down to something like ExecutorsWithContext, ExecutorsWithPrefix or something like that... 

As far as adding additional utility methods, I would like to keep it simple in this change as there is another dependent change. I would rather enrich this on a different change if possible.

Patch-set: 3