Update patch set 4
Patch Set 4:

> There already is a similar class, SpecialExecutors, in yangtools
 > util.concurrent

wasn't aware of that! had a quick look, it's somewhat related, and perhaps Jaime would like to extend the JavaDoc of this one to point to the that one, and the JDK one and the Guava one?  

But SpecialExecutors is new special executors, similar to Guava's MoreExecutors, while this is "just" a wrapper around the existing JDK Executors for the uncaught thread exception thing. Which BTW perhaps you'd like to add to SpecialExecutors, that would probably be useful?

PS: In the longer run, perhaps one day we'd like to one day unify yangtools.util and infrautils.util?  yangtools.util seems to have (almost) nothing to do with yangtools AFAIK, so my preference would be for a move from there to here?  Not urgent at all - I'll merge this here, for the time being..

Patch-set: 4