Update patch set 10
Patch Set 10:

(2 comments)

> > documentation for this feature which I can read
 > > before starting to review the code?

the JavaDoc in the API? There is really only 2 interfaces to look at - the EventPublisher and the EventListener.  Consider that the documentation for this feature, and ignore the rest... ;-)

 > I agree - this needs a write up  as to the intent/purpose. Also
 > this patch is large. I would prefer it broken up into smaller
 > patches with appropriate and detailed commit messages, ie start
 > with an API patch and then introduce the impl in stages.

It's smaller already now as I've removed a number of event types which really are orthogonal to the proposed framework itself, that may help to focus the discussion.  Also note that most LOC really is just a handful of classes forked from Guava (for some tweaking).

Patch-set: 10
Label: Verified=0