RunUntilFailureClassRule with maximumNumberOfTimesToRun
Originally the RunUntilFailureClassRule was intended to find concurrency
issues but only ever used for local testing, but never committed -
because it ran indefinitely.  This made sense for the slower longer
running component tests.

When writing very fast running simple unit tests for concurrency
utilities, it's useful to let them run "a lot of times" even during the
build, to catch potential concurrency issues.

Change-Id: If916b111195e16e1b711716c77b8a3ad97a3e415
Signed-off-by: Michael Vorburger <vorburger@redhat.com>