Bug 54885: Move BP XML from ODL specific to standard location
and remove ODL specific xmlns:odl which are no used in this XML anyway

This is required so that BP errors are detected by SFT in the feature
including this bundle.  The underlying root cause of the problem was
that the respective feature (odl-infrautils-jobcoordinator) does not
depend on the ODL controller feature which installs the blueprint
handler thing, so SFT could never test for issues in this BP XML.

This would require https://git.opendaylight.org/gerrit/#/c/54885/ to be
merged in odlparent, first.  An alternative approach to that (and this)
would be to, somehow, hard-code ALWAYS have the ODL controller feature
which installs the blueprint handler thing, but that seems less modular
to me - I would prefer & recommend that we use this solution instead.

Change-Id: I53a1f046665ef62ed65b91c6dfce15070e2df54f
Signed-off-by: Michael Vorburger <vorburger@redhat.com>