Bug 8415: Ready service integration test (IT), for Karaf 4
This change has an IT passing for Karaf 4 (to confirm that the feature
actually works); another change will get the same to pass under Karaf 3
as well (and that will, likely, require a little bit more work in
odlparent next; see c/56848).

AbstractIntegrationTest now has the (abstract) featureRepositoryURL()
and featureName() methods, which allow ITs to specify the Karaf
features.  It's now also possible to override getKarafURL() to use a
custom distribution, although the idea is that this is typically not
needed.

depends-maven-plugin moved up from itestutils to itestutils-parent.

The alta-maven-plugin is removed, because I've found that at least until
https://ops4j1.jira.com/browse/PAXEXAM-813 is fixed (which I may try to
have a go at, some day) it's not possible to use it with Pax Exam for
Karaf.

ready artifacts are added to infrautils-artifacts POM,
and odl-infrautils-ready to infrautils-features4.

The standard <name> is added to the ready POMs.

Change-Id: I8ad580939bf54c67600a1f06bd052afaa44e68a3
Signed-off-by: Michael Vorburger <vorburger@redhat.com>