Update patch set 2
Patch Set 2:

> > This is failing on the stuff which was introduced to nitrogen
 > > odlparent, which obviously is not available in carbon odlparent.
 > >
 > > Michael, any ideas on how to move switly on this?
 > 
 > Back-patch https://git.opendaylight.org/gerrit/#/c/57650/ from
 > Nitrogen master to 1.8.x-Carbon maintenance in odlparent?  But
 > there may be more issues like this..

We discussed this on the release sync call, and it seems the easiest approach is to drop infrautils from autorelease for now, since it's not needed until we get to Genius. By that point we'll probably have a pre-release of odlparent that infrautils can depend on.

Patch-set: 2
Reviewer: Gerrit User 4540 <4540@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0