Update patch set 2
Patch Set 2:

> Michael, if you consider genius usecases, we have all these async
 > listners and with so many transactions which will get triggered on
 > different nodes in cluster, do you see any way to track things end
 > to end? probably from openstack till ovs ;)

Yeah, but "Rome wasn't built in 1 day" - let's get started somewhere, this kicks off this ball, with something which we can get working (fairly soon, I think) for single node.  Then we'll look at next steps like cluster, and OpenStack and OVS... ;-) Please +1 this if you like the direction this is taking and want to support the move to this! And then, in a few weeks once it works for single node, feel free to ask "and how do we now extend this to even more"... ;)

Patch-set: 2