ExecutionOriginRule new JUnit @Rule
This sets an org.opendaylight.infrautils.utils.mdc.ExecutionOrigin for
each test, and is useful to component test code which would normally run
with such an ExecutionOrigin set into the MDC (in "production"
environment this would be been done e.g. by a TBD Servlet Filter
wrapping a RESTCONF HTTP call, or an ovsdb connection).

This will be required when test run into code using the
CrossThreadContextHolder (proposed in a separate change
https://git.opendaylight.org/gerrit/#/c/61524/), which may expect this
to be present.

It will also be "nice" to see the ExecutionOrigin prefix in logs even
for tests.  This could theoretically also help to easily see in logs any
events from background threads which spill over from one @Test into
another.

Change-Id: Ib03d8fc0e58d938cb0f887a39005536326ec66c3
Signed-off-by: Michael Vorburger <vorburger@redhat.com>