Refactor diagstatus to split out JMX stuff into separate class
I find that this makes the DiagStatusServiceImpl class much shorter and
focused and therefore more easy to understand.  It's really clarifies
the dependency - the DiagStatusServiceMBeanImpl needs a
DiagStatusService - but DiagStatusService(Impl) does not need to know
anything about DiagStatusServiceMBean(Impl).

So this (new) DiagStatusServiceMBeanImpl then just becomes another
consumer of the public DiagStatusService API, just like the CLI, which
IMHO layer this whole thing much more clearly.

This work also revealed something which either is a major design flaw or
possibly a big misunderstanding of the whole thing on my part: There
were a lot of status.getValue() != null checks, which I removed -
because I don't understand how there can ever be an entry in the
statusMap which has no value, for a known key?  This is something
introduced by this change, but it's already like this - this just
reveals that more obviously, because of the clearer separation.

Change-Id: I273b2dd3464519fbe5ad3e6d012f913e3fe4f115
Signed-off-by: Michael Vorburger <vorburger@redhat.com>