Implement FunctionalityReadyNotifierImpl
BundlesBootedReady was already removed
earlier in Ice9cd60878a68037dfcdb384979aabe426cf0a6b.

The BundlesBootedReady was one of my dumber ideas... ;-) because, as
Faseela pointed out in our discussion of diagstatus today, if someone
made some other service dependent on the BundlesBootedReady, e.g. with
an @Inject @OsgiService (or a <service> in BP XML), then that would
create a sort of soft cyclic dependency (not a hard one, SFT would
probably not fail on BP?), because the SystemReadyMonitor would never
reach system state ready ACTIVE - because it would wait on the service
with the dependency to BundlesBootedReady, which means waiting for
SystemReadyMonitor ready, which ... you get the idea.

The *ONLY* valid way to get notified of ready-ness is through the
SystemReadyMonitor registerListener onSystemBootReady().

The FunctionalityReady idea itself still has value though, for any
"higher level" ready-ness signaling; e.g. Daexim's boot-on-import
completed (and I'll put in a change which makes that use this next); but
that is different because it does NOT block

Change-Id: I96960c2250a196ebf6218b7a166aced123202903
Signed-off-by: Michael Vorburger <vorburger@redhat.com>