Remove LoggingRejectedExecutionHandler
On further reflecting around these kind of things tonight, I'm ...
suddenly not sure anymore at all what I was thinking this could possibly
be good for?  I think when originally writing this I was partly inspired
by the LoggingThreadUncaughtExceptionHandler, which IS (very, IMHO)
useful - but this is different...

Got introduced in https://git.opendaylight.org/gerrit/#/c/63350/, BUT:

Executors with unbounded queues will never reject tasks for execution
anyway, will they?  And any which do, such as yangtools bounded
SpecialExecutors, or even the standard JDK ThreadPoolExecutor parent,
seem to have a default RejectedExecutionHandler of new AbortPolicy(),
which throws a RejectedExecutionException.  

So the Executor void execute() caller, would get a
RejectedExecutionException, and have to handle (e.g. log) it.  Or an
ExecutorService submit() would fail with a RejectedExecutionException
(directly, not via a failed Future, according to its JavaDoc). 

So for example in genius' AsyncDataTreeChangeListenerBase, where this
came from, the execute() in onDataTreeChanged would have thrown the
RejectedExecutionException back up to internals of the calling
DataBorker, which, I hope (TBC?) would have logged this themselves?  But
because that usage its with an unbound queue, it could never have
happened there anyway?  And I've just removed that anyway via
https://git.opendaylight.org/gerrit/#/c/63787/ ... ;-)

So... isn't this just stupid?  Even though x3 people +1'd it? ;-)

More than stupid, isn't this outright DANGEROUS - to just LOG but then
silently swallow instead of re-throw - it would make an Executor
[Service] "pretend" to a caller that an execute() or submit() was
successful, and only log, even it wasn't?

If anyone does see any interest in keeping this, please explain why, and
let's update this utility's JavaDoc to explain that better.  And,
probably, make it re-throw???  Else please +1 its removal.

Change-Id: I3cf6fc6ab90d1e1fb7472136975c11b2ec84e2d9
Signed-off-by: Michael Vorburger <vorburger@redhat.com>