Update patch set 3
Patch Set 3:

> I like the general idea, 

Thanks!

 > however I think there’s a bit of an issue
 > with the usage sequencing. Basically, with the way this is used in
 > RollbackCallable 

Duh, right - because you give RollbackCallable your own List, which it wraps for you - but you can still modify your own original behind its back, yeah.. thanks for spotting that. There is really no solution to this!

 > (and would be used in other situations)

Wehl... no, the problem you raise is quite specific to RollbackCallable - its setFutures and getFutures is dumb, one could argue?  If the List was created internally and there was no back-door access (above), then this would be safe?  Let me amend the JavaDoc in RollbackCallable to say a word about this...

 > I think a better approach would be to use Queue instead of List;
 > most implementations of Queue (AFAICT, all apart from LinkedList)
 > reject null.

I don't know - the Queue interface does not (and cannot) guarantee that (as LinkedList shows) - and one not be able to change an existing public API based on List to "just" become a Queue instead.

What to do next here?  How about this: Because Bug 9238 turned out NOT to be related to null in this List, I feel like just removing the RollbackCallable stuff from this change here for now, and let this just be about introducing the NullSafeCollections only, would that work for you?  Without it being used here, yet? ;)

Patch-set: 3
Label: Verified=0