Update patch set 4
Patch Set 4: Code-Review-1

> Duh, right - because you give RollbackCallable your own List, which
 > it wraps for you - but you can still modify your own original
 > behind its back, yeah.. thanks for spotting that. There is really
 > no solution to this!

Yes, there is, and it’s good defensive practice: always copy the contents of collections that are given to you (or that you return).

 > > (and would be used in other situations)
 > 
 > Wehl... no, the problem you raise is quite specific to
 > RollbackCallable - its setFutures and getFutures is dumb, one could
 > argue?  If the List was created internally and there was no
 > back-door access (above), then this would be safe?  Let me amend
 > the JavaDoc in RollbackCallable to say a word about this...

By other situations, I was thinking of all cases where collections are passed around without being copied. Wrapping can never fix incoming collections, it’s only useful for outgoing collections (e.g., wrap your collection in an immutable collection before returning it).

 > > I think a better approach would be to use Queue instead of List;
 > > most implementations of Queue (AFAICT, all apart from LinkedList)
 > > reject null.
 > 
 > I don't know - the Queue interface does not (and cannot) guarantee
 > that (as LinkedList shows) - and one not be able to change an
 > existing public API based on List to "just" become a Queue instead.
 > 
 > What to do next here?  How about this: Because Bug 9238 turned out
 > NOT to be related to null in this List, I feel like just removing
 > the RollbackCallable stuff from this change here for now, and let
 > this just be about introducing the NullSafeCollections only, would
 > that work for you?  Without it being used here, yet? ;)

You know what I think of code added without usage examples...

In any case, I think re. RollbackCallable, that the list should be a hidden implementation detail; and then the nullable checks become part of the implementation of add() and addAll(), either in RollbackCallable itself or delegated to a NullSafe wrapper, in the RollbackCallable use-case that’s neither here nor there.

The big issue really is that presenting this as a null-safe wrapper is risky: if it’s limited to the RollbackCallable use-case (or similar), then it won’t do much, but it gives the impression that it does.  If it’s not limited to that use-case, and aims to “solve” nulls in collections, it would have to check most of its operations, and the semantics become rather complex then. (Collection semantics for equals and hashCode are tough.)

There’s probably a reason Guava doesn’t have anything like this ;-).

Patch-set: 4
Label: Code-Review=-1