RollbackCallable: rework as consumer
The callable’s job here is to act on a single failed job’s results, so
instead of modelling it as a data object, model it as a function
consuming the failed job’s futures and producing its own.

We could just replace RollbackCallable with
Function<List<ListenableFuture<Void>>, List<ListenableFuture<Void>>
but I think keeping the class has its merits, if only for
documentation and readability.

Change-Id: I049f15096afe1c24ecad4e2282f94830fc471398
Signed-off-by: Stephen Kitt <skitt@redhat.com>