Update patch set 4
Patch Set 4:

> Small problem I have is that doing this will break the API, and
 > while RollbackCallable anyway isn't use anywhere currently, I think
 > this will probably interfere with https://git.opendaylight.org/gerrit/#/c/60303/
 > ? But I'm guessing we could bridge/adapt it somehow.  Perhaps I
 > could (finally) finish up c/60303 first and then we get this
 > adjusted so that it doesn't break genius' DJC API, which will
 > delegate to this one?  Or you could propose the same
 > RollbackCallable API change into genius as well, I guess.

AFAICS RollbackCallable, even the Genius version, isn't used anywhere in Genius or NetVirt (except a couple of places in Genius, but they're just forwarding calls so they'd have to change anyway). So I think we can push this in and rework 60303 to use the new API; that won't break anything.

Patch-set: 4
Label: Verified=0