Add JobCoordinator test utils (from genius)
Since the JobCoordinator now resides here and the one in
genius is deprecated, it seems to make sense to move the
associated test utilities here as well, specifically
JobCoordinatorEventsWaiter and TestableJobCoordinatorEventsWaiter.
Also AsyncEventsWaiter since they use that interface. The ones
in genius will be deprecated.

The other option is to leave them in genius and modify
TestableJobCoordinatorEventsWaiter to use JobCoordinator or
create a new version that does.

Change-Id: I9ccd1660c90e266f95fa5711fdfafed429fc318a
Signed-off-by: Tom Pantelis <tompantelis@gmail.com>