Update patch set 3
Patch Set 3:

> The problem with using LogCaptureRule is that some tests purposely
 > test error conditions so ERROR is logged. Not sure if
 > LogCaptureRule can be used with only certain test case(s). 

JUnit rules apply to an entire test class.  My idea was that individual tests who purposely test error conditions would declare this by expectError()

 > there's expectError but that requires a specific error message
 > which looks problematic.

yeah that was just a first stab, I'm very open to any Gerrits proposing to extend that, or even suggestions/emails/JIRAs with idea that I would be willing to extend it with.. what would you need in LogCaptureRule in order to be able to write a test for say this change?  An expectError(Class<Throwable> expectedThrowableClass), kind of thing?

Patch-set: 3
Label: Verified=0