Update patch set 3
Patch Set 3:

> > The problem with using LogCaptureRule is that some tests
 > purposely
 > > test error conditions so ERROR is logged. Not sure if
 > > LogCaptureRule can be used with only certain test case(s).
 > 
 > JUnit rules apply to an entire test class.  My idea was that
 > individual tests who purposely test error conditions would declare
 > this by expectError()
 > 
 > > there's expectError but that requires a specific error message
 > > which looks problematic.
 > 
 > yeah that was just a first stab, I'm very open to any Gerrits
 > proposing to extend that, or even suggestions/emails/JIRAs with
 > idea that I would be willing to extend it with.. what would you
 > need in LogCaptureRule in order to be able to write a test for say
 > this change?  An expectError(Class<Throwable> expectedThrowableClass),
 > kind of thing?

I don't think LogCaptureRule is really suitable for this change.  It could potentially catch a side-effect of attempting to start a task after shutdown. But all that depends on the right timing which you really can't force from a test (w/o introducing latches and thread control in the production code which wold be ugly). So even if there's no ERROR logged, it doesn't mean the   JobQueueHandler thread was actually stopped. We could expose the thread instance for the test to check if it stopped. At the very least, I think we should join with the thread in destroy - I'll change that.

Patch-set: 3
Label: Verified=0