Update patch set 4
Patch Set 4:

I added a join on the JobQueueHandler thread in destroy. I added an accessor so the unit test can verify it's stopped. Ideally the accessor would be only package-scoped but unfortunately the test class is in a separate package so I had to subclass it to get access. I could move the test class to the same package....

Patch-set: 4
Reviewer: Gerrit User 1842 <1842@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0