Update patch set 6
Patch Set 6: Code-Review-1

> We’d said we’d merge this and adapt the DJC-to-infra-JC patches, 
 > so that there would be a single migration...

Oups, sorry; I may have missed a beat / screwed up here?

Anyway, the DJC-to-infra-JC stuff is in now, so would you be willing to raise a new change in genius now to match this change?  Because as-is if I were to merge this now, it would break genius (DataStoreJobCoordinator), so -1 ...

Patch-set: 6
Label: Code-Review=-1