Update patch set 2
Patch Set 2:

> This contradicts with what I'm up to in https://jira.opendaylight.org/browse/ODLPARENT-116
 > (where the Eclipse NullPointer analysis will run on Maven as well),
 > and I would like to find some time to see how we can cater to both
 > FindBugs as well as Eclipse.  BTW: The Eclipse-based null analysis
 > is (far) superior to FindBugs' ...

How is it far superior? I've found FindBugs null analysis to be quite good - it's found issues before and I've played around with it.  It's fine if we can get Eclipse NullPointer analysis  to run and enforce in maven. However we still want to run/enforce FindBugs for all the other checking as well.  

The only thing is being dependent on org.eclipse in production code and being consistent with the rest of ODL that uses the standard javax.annotation. It would be nice if Eclipse supported the standard annotations (IntelliJ does) - is that something they're planning?

Patch-set: 2