Update patch set 1
Patch Set 1:

NeutronFloatingToFixedIpMappingChangeListener - it's a DTCL so it's notified synchronously and doesn't need locking. There's one outlier call from the outside. NeutronvpnManager is the other class that uses the locking. I really don't know what the author's intent was  - the patch that added it says nothing about it.  The common denominator is that it's locking concurrent access to the DS for an entity identified by a nebulous lock name. But why.... the DS does optimistic locking, the lock is really a tryLock so it may get the lock after another operation completes.  The only thing I can think of is they're trying to prevent optimistic lock failures. Of course there shouldn't be multiple writers anyway. 

Im' inclined to submit a patch to remove the locking and see what folks say.

Patch-set: 1