Update patch set 1
Patch Set 1:

In looking at NeutronUtils.lock, it does:

   writeLock().tryLock(LOCK_WAIT_TIME, secUnit);

but ignores and doesn't return the value from tryLock 
(FB flags this).  if tryLock returns false that means it timed out and didn't get the lock. But NeutronUtils.lock falls thru and always returns true. It never return false - it would throw a RuntimeException if interrupted which isn't good either. Even if it did return false, callers don't check it and still proceed. Even worse, they only use the return value to determine whether to unlock. So if tryLock returned false, the caller would see true and and call unlock which would throw an exception  b/c the current thread doesn't own the lock. 

So the code only works right if there's no contention or tryLock is able to get the lock before timing out.

Patch-set: 1