Metrics Step 3: jobcoordinator using metrics instead of counters
With this change, the jobcoordinator stats are now available via JMX,
and can be periodically dumped to files with the upcoming
I033aa8628de8bfee8cdca7da564eb45a8c567c52.

This is strictly speaking more of a bug fix than a new feature, because
DJC originally used to be monitor-able with infrautils.counters, but we
had to remove that; this bring that back (differently); also see
INFRAUTILS-19 for some more background.

PS: I'd like to eventually remove infrautils.counters completely.

Change-Id: I8620996d3be56d86c85f184057e1a23ae3119140
Signed-off-by: Michael Vorburger <vorburger@redhat.com>