Update patch set 4
Patch Set 4:

Viji, thanks for this.  If you have time to have a closer look at my code starting to coming in on the "metrics" topic, IMHO that would be the best way to align the requirements you are outlining here with that work.  In an ideal world, if you could e.g. switch genius arputils' use of infrautils.counters to infrautils.metrics, that would be a great way to familiarize yourself with it.

Patch-set: 4
Reviewer: Gerrit User 5867 <5867@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0