Update patch set 1
Patch Set 1:

> Change seems fine - just curious if there was any issue with using
 > odlparent or is it to use the lite version b/c we don't need the
 > "heavy" version?

yeah, just that.  If have a separate odlparent VS odlparent-lite, then aggregrators should use it.. most of them do, I just stumbled upon this one not doing it right.  It leads to a minor build speed up.  (One could argue that a number of projects probably don't get this, so that perhaps it would be simpler to have only one parent, perhaps some smarts based on Maven profiles to skip stuff if no sources? But it's probably way too late for that, so let us at least use what we have correctly.)

Patch-set: 1