Metrics new API, decoupled from Dropwizard Codahale's
because I'm hoping, some fine day in 2019 when I have time (never?), to
provide an alternative implementation backed directly & natively by
http://Prometheus.io, as an alternative to Dropwizard Codahale's.

The new API mimics the original one closely (and it's trivial to switch
any code already using it..), minus any kind of "reporting" methods such
as getFiveMinuteRate() etc. That is fine because those methods should be
of no interest to applications who "provide" metrics - only to those who
want to report - which in this design should just be metrics.impl, via
Dropwizard's JMX and File etc. reporters, and in the future Prometheus.

While we're at it, we make the Meter, Counter & Timer close-able for
unregistration, and allow Timer to time using the Checked Callable and
Runnable, which is handy.

Change-Id: I76ce2e1600e8f99eb735aac4c29c423cc01c59ee
Signed-off-by: Michael Vorburger <vorburger@redhat.com>