Update patch set 2
Patch Set 2:

> Can MetricProvider API have one more API to remove the metric from
 > the registry! Certain Applications need to control the lifecycle of
 > the metrics at runtime.Codahale also has an API for it.
 > public boolean removeMetric(Object anchor, String id) {
 > return registry.remove(id);
 > }

but that is what the close() on the Meter etc. are for! Did you see that?

 > Michael, Is there a plan to implement guages sooner. If so, when
 > can we expect? We have applications to be migrated which will be
 > using guages.

I was just lazy ;) do you have 1h to raise a follow up patch which adds that? It should be trivial, I think.

If this is OK for you, could you +1 it so we can merge it?

Patch-set: 2