add get() method to metrics Counter and Meter
because I realized while adapting the JobCoordinator to metrics in
I8620996d3be56d86c85f184057e1a23ae3119140 that for tests a get() does
have its uses, after all... ;-)

But it's really intended just for tests, and not other uses. So no plans
to, and IMHO in this API it would not make sense, to expose more like
e.g. via methods such as Dropwizard's getMeanRate() or
getFifteenMinuteRate() ... those are implementation details, which
application clients should never require - and with a future Prometheus
or Ceilometer instead of Dropwizard implementation the Java client of
such system would possibly not even expose that, as such aggregation
would happen inside such external metrics servers.

Timer intentionally has no get() as tests should never be needing that;
and it would be totally unreliable to test for that, anyway.

Change-Id: Ic02a483d2c33c3279369ccfa98362c6ad50892bc
Signed-off-by: Michael Vorburger <vorburger@redhat.com>