Update patch set 3
Patch Set 3:

This patch intends to provide a uniform naming format which is good. Based on your comments in other patches, I got the impression this patch is also intended to avoid users doing: 

  Map<String, Counter>  counterMap;

If that is correct then the proposed Labeled interface always creates a new instance and does not store instances so users will still need to maintain a Map. This is OK but I'm wondering if we could also provide that functionality here, since it appears to be a common pattern,  with a LabeledMetricCache (or Map or something else - I know you don't like term "Cache" here :))

Patch-set: 3