Update patch set 5
Patch Set 5:

> Perhaps we're over-thinking/over-engineering this - trying to
 > anticipate someone wanting 1, 2 ,3  labels (what if someone wants
 > 4?) and what format. 

then that someone can follow the pattern used here and contribute another method with 4.  If you have an idea how to design such an API in a better way yet keep it strongly statically typed, I'd love to see how...

 > How about we just keep it simple and flexible
 > and just allow a format string and args in Labeled:
 > 
 > T label(String labelFormat, Object... args);

Huh, now how would this be used?  These labels have to be pairs of name/values, to be able to pass them to metrics systems with support for labels, such as https://Prometheus.io, where there is an API to add such pairs of name/values, and then exploit that to do aggregation functions and such over such labels.  You seem to be proposing an arbitrary log message like format & args... that's not what's needed, here; I think the fact that the implementation for Codahale's Dropwizard Metrics back-end turns them into a String is perhaps confusing?

Patch-set: 5